apps/backend/app/research_agent.py
import os
import asyncio
from typing import Dict, Any
from dotenv import load_dotenv

load_dotenv()

class ResearchAgent:
    def __init__(self):
        # Simple initialization
        pass
    
    async def research_async(self, query: str, max_sources: int = 5, detailed: bool = True) -> Dict[str, Any]:
        """Perform research on a given query"""
        
        # Simulate research time
        await asyncio.sleep(2)
        
        # Return comprehensive mock research results
        return {
            "query": query,
            "content": f"""# Research Report: {query}

## Executive Summary
This comprehensive analysis examines the current state and future prospects of {query}. Based on extensive research across {max_sources} authoritative sources, this report provides key insights and actionable recommendations.

## Key Findings

### 1. Current Landscape
- Growing adoption and integration of {query} across industries
- Emerging trends in {query} implementation and optimization
- Significant market opportunities in the {query} space

### 2. Technical Analysis
- Core technologies enabling {query} development
- Performance considerations and scalability factors
- Integration challenges and solutions

### 3. Market Impact
- Projected growth rates for {query} through 2025
- Key players and competitive landscape
- Regional adoption patterns and regulatory considerations

## Recommendations

1. **Short-term**: Focus on foundational {query} capabilities
2. **Medium-term**: Develop specialized {query} applications  
3. **Long-term**: Establish thought leadership in {query} innovation

## Methodology
This analysis synthesized data from {max_sources} sources including:
- Peer-reviewed academic publications
- Industry market reports
- Expert interviews and case studies
- Technical documentation and benchmarks

*Note: This is a mock research report. Real AI-powered research capabilities will be integrated in the next development phase.*""",
            "sources_used": max_sources,
            "detailed": detailed,
            "status": "completed"
        }