apps/backend/pyproject.toml
[build-system]
requires = ["flit_core>=3.2"]
build-backend = "flit_core.buildapi"

[project]
name = "research-agent-backend"
version = "0.1.0"
description = "Backend for AI Research Agent"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "langchain>=0.0.353",
    "langchain-community>=0.0.10",
    "langgraph>=0.0.40",
    "openai>=1.3.0",
    "groq>=0.3.0",
    "python-multipart>=0.0.6",
    "supabase>=1.0.3",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[tool.pyright]
include = ["app"]
extraPaths = ["app"]